% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{fix_datetime}
\alias{fix_datetime}
\title{fix inconsistently encoded datetime
`can_fix <- c("D0 _13:48:51", "D0_13:48:51", "13:48:51")``}
\usage{
fix_datetime(v, start_date)
}
\arguments{
\item{v}{vector of strings to fix}

\item{start_date}{start date that corresponds to D0 (day 0).}
}
\description{
fix inconsistently encoded datetime
`can_fix <- c("D0 _13:48:51", "D0_13:48:51", "13:48:51")``
}
\keyword{internal}
