% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{fix_temperature}
\alias{fix_temperature}
\title{FIXME: turn these into actual test cases:
can_fix <- c('300', '30.1')}
\usage{
fix_temperature(v)
}
\description{
FIXME: turn these into actual test cases:
can_fix <- c('300', '30.1')
}
