% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_dims_from_n}
\alias{guess_dims_from_n}
\title{guess the dimensions of microtitre plate from the number of wells.}
\usage{
guess_dims_from_n(n)
}
\arguments{
\item{n}{\code{<num>} number of wells}
}
\value{
\code{\link[=list]{list()}} of length 2: number of rows (\code{nrow}) and columns (\code{ncol}).
}
\description{
guess the dimensions of microtitre plate from the number of wells.
}
\keyword{internal}
