% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{parse_section_data}
\alias{parse_section_data}
\title{read the data sections from ini yg files}
\usage{
parse_section_data(l, start_date)
}
\arguments{
\item{l}{list of parsed ini file}

\item{start_date}{date that signifies 'D0' in the raw text file. Because the
format encodes only relative datetime in the data section need to have
standard date be supplied externally.}
}
\description{
read the data sections from ini yg files
}
