% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ini.R
\name{read_ini}
\alias{read_ini}
\title{Read ini file}
\usage{
read_ini(path, omit = NULL, encoding = "UTF-8")
}
\arguments{
\item{path}{<chr> path to file}

\item{omit}{<chr> section name(s) to omit}

\item{encoding}{<chr> defaults to 'UTF-8' encoding.}
}
\description{
Compared to \code{ini::read.ini()} is ~4x faster and allows you to omit
certain sections.
}
\details{
This function is used internally to parse the yeast grower data ini files.
}
