% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ini.R
\name{section_key_valuefy}
\alias{section_key_valuefy}
\title{take a list generated by an ini file and turn it into a}
\usage{
section_key_valuefy(l)
}
\description{
take a list generated by an ini file and turn it into a
}
