% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_labels}
\alias{well_labels}
\title{well labels from dimensions of microtitre plate.}
\usage{
well_labels(nrow, ncol, col_nchar = nchar(ncol), order = c("rowmajor",
  "colmajor"))
}
\arguments{
\item{nrow}{\code{<int>} number of rows}

\item{ncol}{\code{<int>} number of cols}

\item{col_nchar}{\code{<int>} the number of characters for the column section. This
controls the left-padding of well labels with zeros. e.g. A1 vs. A01. The
default is the number of characters in \code{ncol} for nice lexicographic
sorting behaviour.}

\item{order}{\code{<chr>} defaults to \code{"rowmajor"}, alternatively \code{"colmajor"}}
}
\value{
\code{<chr>} vector of well labels where the row is represented by a letter,
and the col is represented by an intenger, possibly left-padded with one
or more zeros.
}
\description{
well labels from dimensions of microtitre plate.
}
\keyword{internal}
