% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_labels_from_length}
\alias{well_labels_from_length}
\title{get the labels for a microtitre plate based on the length of the vector.}
\usage{
well_labels_from_length(v)
}
\arguments{
\item{v}{vector of length on which guess is based.}
}
\description{
get the labels for a microtitre plate based on the length of the vector.
}
\keyword{internal}
