% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yg-example.R
\name{yg_example_files}
\alias{yg_example_files}
\title{get file paths to example yeast grower files provided with the package.}
\usage{
yg_example_files()
}
\description{
get file paths to example yeast grower files provided with the package.
}
