% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{yg_read_data}
\alias{yg_read_data}
\title{read the data sections from ini yg files}
\usage{
yg_read_data(l, path)
}
\arguments{
\item{l}{list of parsed ini file}
}
\description{
read the data sections from ini yg files
}
