% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{yg_read_info}
\alias{yg_read_info}
\title{read the info sections from ini yg files}
\usage{
yg_read_info(l)
}
\arguments{
\item{l}{list of parsed ini file}
}
\description{
read the info sections from ini yg files
}
