% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{fix_datetime}
\alias{fix_datetime}
\title{FIXME: turn these into actual test cases:
can_fix <- c("D0 _13:48:51", "D0_13:48:51", "13:48:51")}
\usage{
fix_datetime(v, start_date)
}
\description{
FIXME: turn these into actual test cases:
can_fix <- c("D0 _13:48:51", "D0_13:48:51", "13:48:51")
}
