% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-yg.R
\name{fix_temperature}
\alias{fix_temperature}
\title{fix inconsistently encoded temperatures: \code{'300'} and \code{'30.0'} made consistent.}
\usage{
fix_temperature(v)
}
\arguments{
\item{v}{vector of strings to fix}
}
\description{
fix inconsistently encoded temperatures: \code{'300'} and \code{'30.0'} made consistent.
}
\keyword{internal}
