% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ini.R
\name{read_ini}
\alias{read_ini}
\title{Read ini formatted text file into a nested section-key-value list.}
\usage{
read_ini(path)
}
\arguments{
\item{path}{<chr> path to file}
}
\description{
This is re-written from the \code{ini::read.ini()} function to use stringr for
tokenization and matching. This simplifies the source code and makes it
~3 times faster. It is also strictly reading in UTF-8.
}
\details{
This function is used to parse the yeast grower data.

\code{identical(read_ini(path), read.ini(path))} returns TRUE.
}
